+++ package.json
@@ -24,7 +24,6 @@
     "check:boundaries": "bun run scripts/check-monorepo-boundaries.ts",
     "check:api-validation": "bun run scripts/check-api-validation-contracts.ts --check",
     "check:api-validation:strict": "bun run scripts/check-api-validation-contracts.ts --check --enforce-boundary-baseline",
-    "check:block-canonical": "bun run apps/sim/scripts/check-block-canonical.ts",
     "check:realtime-prune": "bun run scripts/check-realtime-prune-graph.ts",
     "mship-contracts:generate": "bun run scripts/sync-mothership-stream-contract.ts",
     "mship-contracts:check": "bun run scripts/sync-mothership-stream-contract.ts --check",
